Added scenario progression trigger for git push.
